Add karafka and avro gems. Add account consumer in analytics. Fix account consumer in another services.

+++ analytics/app/consumers/account_consumer.rb
@@ -1,7 +1,6 @@
 # frozen_string_literal: true
 
-# Example consumer that prints messages payloads
-class AccountChangesConsumer < ApplicationConsumer
+class AccountConsumer < ApplicationConsumer
   def consume
     messages.each do |message|
       payload = AVRO.decode(message.raw_payload)
@@ -12,18 +11,20 @@ class AccountChangesConsumer < ApplicationConsumer
 
       case payload["event_name"]
       when "AccountCreated"
-        Account.create(email: payload["data"]["email"],
-                       full_name: payload["data"]["full_name"],
-                       public_id: payload["data"]["public_id"],
-                       role: payload["data"]["role"])
+        Account.find_or_create_by(public_id: payload["data"]["public_id"]).tap do |acc|
+          acc.email = payload["data"]["email"] unless acc.email
+          acc.full_name = payload["data"]["full_name"] unless acc.full_name
+          acc.role = payload["data"]["role"] unless acc.role
+          acc.save!
+        end
       when "AccountUpdated"
-        account = Account.find_by(public_id: payload["data"]["public_id"])
+        account = Account.find_or_create_by(public_id: payload["data"]["public_id"])
         account.update(email: payload["data"]["email"],
                        full_name: payload["data"]["full_name"],
                        role: payload["data"]["role"])
       when "AccountDeleted"
         account = Account.find_by(public_id: payload["data"]["public_id"])
-        account.destroy
+        account.destroy if account
       end
     end
   end
